refactor: reorganize imports and move resource definitions to resources.py
